+++ CHANGELOG.md
@@ -22,7 +22,7 @@
 **Terraform**
 - Always return value for `plan`, `test`, and `validate` methods.
 - Return `Boolean` for `fmt`.
-- Add `compactWarn` parameter to `apply` method.
+- Add `compactWarn` parameter to `apply` and `destroy` methods.
 
 ### 2.1.3
 **AWX**